AutoMapper and FluentValidation

+++ OnionPronia/src/Core/OnionPronia.Domain/Entities/Product.cs
@@ -8,11 +8,12 @@ namespace OnionPronia.Domain.Entities
 {
     public class Product:BaseNameableEntity
     {
+
         public decimal Price { get; set; }
         public string SKU { get; set; }
         public string Description { get; set; }
         //relational
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         public Category Category { get; set; }
         public ICollection<ProductTag> ProductTags { get; set; }
